pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cfs-routes"
version = "0.1.0"
description = "CFS Preferred IFR Routes service for VATSIM controllers"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "alembic",
    "pdfplumber",
    "httpx",
    "apscheduler~=3.10",
    "aiomysql",
    "aiosqlite",
    "pydantic-settings",
    "jinja2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]
